Renamed handler to settings
