+++ core/frontend/services/redirects/settings.js
