+++ core/server/api/v0.1/redirects.js
@@ -8,12 +8,12 @@ redirectsAPI = {
     download(options) {
         return localUtils.handlePermissions('redirects', 'download')(options)
             .then(() => {
-                return redirects.handler.serveDefault();
+                return redirects.settings.serveDefault();
             });
     },
     upload(options) {
         return localUtils.handlePermissions('redirects', 'upload')(options)
-            .then(() => redirects.handler.activate(options.path))
+            .then(() => redirects.settings.activate(options.path))
             .then(() => {
                 // CASE: trigger that redirects are getting re-registered
                 web.shared.middlewares.customRedirects.reload();
